Extract Auth test in another file, fix IT
