+++ src/test/kotlin/com/github/imflog/schema/registry/compatibility/CompatibilityTaskIT.kt
@@ -1,102 +1,40 @@
 package com.github.imflog.schema.registry.compatibility
 
-import com.fasterxml.jackson.databind.ObjectMapper
-import com.github.imflog.schema.registry.REGISTRY_FAKE_AUTH_PORT
 import com.github.imflog.schema.registry.TestContainersUtils
-import com.github.tomakehurst.wiremock.WireMockServer
-import com.github.tomakehurst.wiremock.client.WireMock
-import com.github.tomakehurst.wiremock.common.ConsoleNotifier
-import com.github.tomakehurst.wiremock.core.WireMockConfiguration
 import io.confluent.kafka.schemaregistry.avro.AvroSchema
-import io.confluent.kafka.schemaregistry.client.rest.entities.Schema
 import org.assertj.core.api.Assertions
 import org.gradle.internal.impldep.org.junit.rules.TemporaryFolder
 import org.gradle.testkit.runner.BuildResult
 import org.gradle.testkit.runner.GradleRunner
 import org.gradle.testkit.runner.TaskOutcome
-import org.junit.jupiter.api.AfterAll
 import org.junit.jupiter.api.AfterEach
-import org.junit.jupiter.api.BeforeAll
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import java.io.File
 
-class CompatibilityTaskIT: TestContainersUtils() {
-    lateinit var folderRule: TemporaryFolder
-    lateinit var buildFile: File
+class CompatibilityTaskIT : TestContainersUtils() {
+    private lateinit var folderRule: TemporaryFolder
+    private lateinit var buildFile: File
 
-    private val username: String = "user"
-
-    private val password: String = "pass"
+    private val subject = "test-subject"
 
     private val defaultSchema = """{
-        "type":"record",
-        "name":"Dependency",
+        "type": "record",
+        "name": "User",
         "fields":[
             {
-                "name":"name",
-                "type":"string"
+                "name": "name",
+                "type": "string"
             }
         ]
     }"""
-
-    // Refacto this when switching to TestContainers.
-    private val mapper = ObjectMapper()
-
-    companion object {
-        lateinit var wireMockAuthServerItem: WireMockServer
-
-        @BeforeAll
-        @JvmStatic
-        fun initClass() {
-            wireMockAuthServerItem = WireMockServer(
-                WireMockConfiguration
-                    .wireMockConfig()
-                    .port(REGISTRY_FAKE_AUTH_PORT)
-                    .notifier(ConsoleNotifier(true))
-            )
-            wireMockAuthServerItem.start()
-        }
-
-        @AfterAll
-        @JvmStatic
-        fun tearDown() {
-            wireMockAuthServerItem.stop()
-        }
-    }
 
     @BeforeEach
     fun init() {
         // Reset the client before each test
         folderRule = TemporaryFolder()
 
-        client.register("Dependency", AvroSchema(defaultSchema))
-
-        // Stub with authentication configuration
-        wireMockAuthServerItem.stubFor(
-            WireMock.post(
-                WireMock
-                    .urlMatching("/compatibility/subjects/.*/versions/.*")
-            )
-                .withBasicAuth(username, password)
-                .willReturn(
-                    WireMock.aResponse()
-                        .withStatus(200)
-                        .withHeader("Content-Type", "application/vnd.schemaregistry.v1+json")
-                        .withBody("{\"is_compatible\": true}")
-                )
-        )
-
-        val schema = Schema("Dependency", 1, 1, AvroSchema.TYPE, listOf(), defaultSchema)
-        wireMockAuthServerItem.stubFor(
-            WireMock
-                .get(WireMock.urlMatching("/subjects/.*/versions/1.*"))
-                .willReturn(
-                    WireMock.aResponse()
-                        .withStatus(200)
-                        .withBody(mapper.writeValueAsString(schema))
-                )
-        )
+        client.register(subject, AvroSchema(defaultSchema))
     }
 
     @AfterEach
@@ -108,23 +46,25 @@ class CompatibilityTaskIT: TestContainersUtils() {
     fun `CompatibilityTask should validate input schema with no dependencies`() {
         folderRule.create()
         folderRule.newFolder("avro")
-        val testAvsc = folderRule.newFile("avro/other_test.avsc")
+        val testAvsc = folderRule.newFile("avro/test.avsc")
         val schemaTest = """
             {
-                "type":"record",
-                "name":"Blah",
-                "fields":[
+                "type": "record",
+                "name": "User",
+                "fields": [
                     {
-                        "name":"name",
-                        "type":"string"
+                        "name": "name",
+                        "type": "string"
+                    },
+                    {
+                        "name": "nickname",
+                        "type": ["null", "string"],
+                        "default": null
                     }
                 ]
             }
-        """.trimIndent()
+        """
         testAvsc.writeText(schemaTest)
-
-        val testAvsc2 = folderRule.newFile("avro/test.avsc")
-        testAvsc2.writeText(schemaTest)
 
         buildFile = folderRule.newFile("build.gradle")
         buildFile.writeText(
@@ -135,14 +75,9 @@ class CompatibilityTaskIT: TestContainersUtils() {
             }
 
             schemaRegistry {
-                url = 'http://localhost:$REGISTRY_FAKE_AUTH_PORT/'
-                credentials {
-                    username = '$username'
-                    password = '$password'
-                }
+                url = '$schemaRegistryEndpoint'
                 compatibility {
-                    subject('testSubject1', 'avro/test.avsc')
-                    subject('testSubject2', 'avro/other_test.avsc')
+                    subject('$subject', 'avro/test.avsc')
                 }
             }
         """
@@ -164,127 +99,6 @@ class CompatibilityTaskIT: TestContainersUtils() {
         buildFile = folderRule.newFile("build.gradle")
         buildFile.writeText(
             """
-            import io.confluent.kafka.schemaregistry.client.rest.entities.SchemaReference
-
-            plugins {
-                id 'java'
-                id 'com.github.imflog.kafka-schema-registry-gradle-plugin'
-            }
-
-            schemaRegistry {
-                url = 'http://localhost:$REGISTRY_FAKE_AUTH_PORT/'
-                credentials {
-                    username = '$username'
-                    password = '$password'
-                }
-                compatibility {
-                    subject('testSubject', 'avro/core.avsc', "AVRO").addReference('Dependency', 'Dependency', 1)
-                }
-            }
-        """
-        )
-
-        folderRule.newFolder("avro")
-        val coreAvsc = folderRule.newFile("avro/core.avsc")
-        val coreSchema = """
-            {
-                "type":"record",
-                "name":"Core",
-                "fields":[
-                    {
-                        "name":"dep",
-                        "type":"Dependency"
-                    }
-                ]
-            }
-        """.trimIndent()
-        coreAvsc.writeText(coreSchema)
-
-        val result: BuildResult? = GradleRunner.create()
-            .withGradleVersion("6.2.2")
-            .withProjectDir(folderRule.root)
-            .withArguments(CompatibilityTask.TASK_NAME)
-            .withPluginClasspath()
-            .withDebug(true)
-            .build()
-        Assertions.assertThat(result?.task(":testSchemasTask")?.outcome).isEqualTo(TaskOutcome.SUCCESS)
-    }
-
-    @Test
-    internal fun `Compatibility task should fail if credential not provided on authenticated registry`() {
-        folderRule.create()
-        folderRule.newFolder("avro")
-        val testAvsc = folderRule.newFile("avro/other_test.avsc")
-        val schemaTest = """
-            {
-                "type":"record",
-                "name":"Blah",
-                "fields":[
-                    {
-                        "name":"name",
-                        "type":"string"
-                    }
-                ]
-            }
-        """.trimIndent()
-        testAvsc.writeText(schemaTest)
-
-        val testAvsc2 = folderRule.newFile("avro/test.avsc")
-        testAvsc2.writeText(schemaTest)
-
-        buildFile = folderRule.newFile("build.gradle")
-        // Auth configuration is omited to test the failure
-        buildFile.writeText(
-            """
-            plugins {
-                id 'java'
-                id 'com.github.imflog.kafka-schema-registry-gradle-plugin'
-            }
-
-            schemaRegistry {
-                url = 'http://localhost:$REGISTRY_FAKE_AUTH_PORT/'
-                compatibility {
-                    subject('testSubject1', 'avro/test.avsc')
-                }
-            }
-        """
-        )
-
-        val result: BuildResult? = GradleRunner.create()
-            .withGradleVersion("6.2.2")
-            .withProjectDir(folderRule.root)
-            .withArguments(CompatibilityTask.TASK_NAME)
-            .withPluginClasspath()
-            .withDebug(true)
-            .buildAndFail()
-        Assertions.assertThat(result?.task(":testSchemasTask")?.outcome).isEqualTo(TaskOutcome.FAILED)
-    }
-
-    @Test
-    fun `CompatibilityTask should validate input schema with dependencies without credentials`() {
-        folderRule.create()
-        folderRule.newFolder("avro")
-        val testAvsc = folderRule.newFile("avro/other_test.avsc")
-        val schemaTest = """
-            {
-                "type":"record",
-                "name":"Blah",
-                "fields":[
-                    {
-                        "name":"name",
-                        "type":"string"
-                    }
-                ]
-            }
-        """.trimIndent()
-        testAvsc.writeText(schemaTest)
-
-        val testAvsc2 = folderRule.newFile("avro/test.avsc")
-        testAvsc2.writeText(schemaTest)
-
-        buildFile = folderRule.newFile("build.gradle")
-        buildFile.writeText(
-            """
             plugins {
                 id 'java'
                 id 'com.github.imflog.kafka-schema-registry-gradle-plugin'
@@ -292,17 +106,46 @@ class CompatibilityTaskIT: TestContainersUtils() {
 
             schemaRegistry {
                 url = '$schemaRegistryEndpoint'
-                credentials {
-                    username = '$username'
-                    password = '$password'
-                }
                 compatibility {
-                    subject('testSubject1', 'avro/test.avsc')
-                    subject('testSubject2', 'avro/other_test.avsc')
+                    subject('$subject', 'avro/test.avsc', "AVRO").addReference('Address', 'Address', 1)
                 }
             }
         """
         )
+
+        client.register("Address", AvroSchema("""
+            {
+                "type": "record",
+                "name": "Address",
+                "fields": [
+                    {
+                        "name": "street",
+                        "type": "string"
+                    }
+                ]
+            }
+        """))
+
+        folderRule.newFolder("avro")
+        val testAvsc = folderRule.newFile("avro/test.avsc")
+        val schemaTest = """
+            {
+                "type":"record",
+                "name":"User",
+                "fields":[
+                    {
+                        "name": "name",
+                        "type": "string"
+                    },
+                    {
+                        "name": "address",
+                        "type": ["null", "Address"],
+                        "default": null
+                    }
+                ]
+            }
+        """
+        testAvsc.writeText(schemaTest)
 
         val result: BuildResult? = GradleRunner.create()
             .withGradleVersion("6.2.2")
