+++ src/test/kotlin/com/github/imflog/schema/registry/download/DownloadTaskIT.kt
@@ -1,22 +1,13 @@
 package com.github.imflog.schema.registry.download
 
-import com.fasterxml.jackson.databind.ObjectMapper
-import com.github.imflog.schema.registry.REGISTRY_FAKE_AUTH_PORT
 import com.github.imflog.schema.registry.TestContainersUtils
-import com.github.tomakehurst.wiremock.WireMockServer
-import com.github.tomakehurst.wiremock.client.WireMock
-import com.github.tomakehurst.wiremock.common.ConsoleNotifier
-import com.github.tomakehurst.wiremock.core.WireMockConfiguration
 import io.confluent.kafka.schemaregistry.avro.AvroSchema
-import io.confluent.kafka.schemaregistry.client.rest.entities.Schema
 import org.assertj.core.api.Assertions
 import org.gradle.internal.impldep.org.junit.rules.TemporaryFolder
 import org.gradle.testkit.runner.BuildResult
 import org.gradle.testkit.runner.GradleRunner
 import org.gradle.testkit.runner.TaskOutcome
-import org.junit.jupiter.api.AfterAll
 import org.junit.jupiter.api.AfterEach
-import org.junit.jupiter.api.BeforeAll
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import java.io.File
@@ -26,69 +17,37 @@ class DownloadTaskIT : TestContainersUtils() {
     private lateinit var folderRule: TemporaryFolder
     private val subject = "test-subject"
 
-    private val username: String = "user"
+    private val schemaOld = """{
+        "type": "record",
+        "name": "User",
+        "fields": [
+            { "name": "name", "type": "string" }
+        ]
+    }"""
 
-    private val password: String = "pass"
-
-    val schema =
-        "{\"type\": \"record\", \"name\": \"Blah\", \"fields\": [{ \"name\": \"name\", \"type\": \"string\" }]}"
-
-    private val schemaOld =
-        "{\"type\": \"record\", \"name\": \"Blah\", \"fields\": [{ \"name\": \"name\", \"type\": \"string\" }, { \"name\": \"description\", \"type\": \"string\" }]}"
+    private val schema = """{
+        "type": "record",
+        "name": "User",
+        "fields": [
+            { "name": "name", "type": "string" }, 
+            { "name": "description", "type": ["null", "string"], "default": null }
+        ]
+    }"""
 
     private lateinit var buildFile: File
 
-    private val mapper = ObjectMapper()
-
-    companion object {
-        lateinit var wiremockAuthServerItem: WireMockServer
-
-        @BeforeAll
-        @JvmStatic
-        fun initClass() {
-            wiremockAuthServerItem = WireMockServer(
-                WireMockConfiguration
-                    .wireMockConfig()
-                    .port(REGISTRY_FAKE_AUTH_PORT)
-                    .notifier(ConsoleNotifier(true))
-            )
-            wiremockAuthServerItem.start()
-        }
-
-        @AfterAll
-        @JvmStatic
-        fun tearDown() {
-            wiremockAuthServerItem.stop()
-        }
-    }
-
     @BeforeEach
     fun init() {
-        folderRule = TemporaryFolder()
-        folderRule.create()
-
         // Register schema
         client.register(subject, AvroSchema(schemaOld))
         client.register(subject, AvroSchema(schema))
 
-        // Stub with authentication configuration
-        wiremockAuthServerItem.stubFor(
-            WireMock.get(
-                WireMock
-                    .urlMatching("/subjects/test-subject/versions/latest")
-            )
-                .withBasicAuth(username, password)
-                .willReturn(
-                    WireMock.aResponse()
-                        .withStatus(200)
-                        .withHeader("Accept", "application/json")
-                        .withBody(mapper.writeValueAsString(Schema(subject, 1, 1, "AVRO", emptyList(), schema)))
-                )
-        )
+        folderRule = TemporaryFolder()
+        folderRule.create()
     }
 
     @AfterEach
-    internal fun tearDown() {
+    fun tearDown() {
         folderRule.delete()
     }
 
@@ -103,45 +62,9 @@ class DownloadTaskIT : TestContainersUtils() {
             }
 
             schemaRegistry {
-                url = 'http://localhost:$REGISTRY_FAKE_AUTH_PORT/'
-                credentials {
-                    username = '$username'
-                    password = '$password'
-                }
-                download {
-                    subject('test-subject', 'src/main/avro/test')
-                }
-            }
-        """
-        )
-
-        val result: BuildResult? = GradleRunner.create()
-            .withGradleVersion("6.2.2")
-            .withProjectDir(folderRule.root)
-            .withArguments(DownloadTask.TASK_NAME)
-            .withPluginClasspath()
-            .withDebug(true)
-            .build()
-
-        Assertions.assertThat(File(folderRule.root, "src/main/avro/test")).exists()
-        Assertions.assertThat(File(folderRule.root, "src/main/avro/test/test-subject.avsc")).exists()
-        Assertions.assertThat(result?.task(":downloadSchemasTask")?.outcome).isEqualTo(TaskOutcome.SUCCESS)
-    }
-
-    @Test
-    fun `DownloadSchemaTask should download last schema version without credentials`() {
-        buildFile = folderRule.newFile("build.gradle")
-        buildFile.writeText(
-            """
-            plugins {
-                id 'java'
-                id 'com.github.imflog.kafka-schema-registry-gradle-plugin'
-            }
-
-            schemaRegistry {
                 url = '$schemaRegistryEndpoint'
                 download {
-                    subject('test-subject', 'src/main/avro/test')
+                    subject('$subject', 'src/main/avro/test')
                 }
             }
         """
@@ -190,36 +113,7 @@ class DownloadTaskIT : TestContainersUtils() {
     }
 
     @Test
-    fun `DownloadSchemaTask should fail download when credentials not setup and required`() {
-        buildFile = folderRule.newFile("build.gradle")
-        buildFile.writeText(
-            """
-            plugins {
-                id 'java'
-                id 'com.github.imflog.kafka-schema-registry-gradle-plugin'
-            }
-
-            schemaRegistry {
-                url = 'http://localhost:$REGISTRY_FAKE_AUTH_PORT/'
-                download {
-                    subject('test-subject', 'src/main/avro/test')
-                }
-            }
-        """
-        )
-
-        val result: BuildResult? = GradleRunner.create()
-            .withGradleVersion("6.2.2")
-            .withProjectDir(folderRule.root)
-            .withArguments(DownloadTask.TASK_NAME)
-            .withPluginClasspath()
-            .withDebug(true)
-            .buildAndFail()
-        Assertions.assertThat(result?.task(":downloadSchemasTask")?.outcome).isEqualTo(TaskOutcome.FAILED)
-    }
-
-    @Test
-    internal fun `DownloadSchemaTask should download specific schema version without credentials`() {
+    fun `DownloadSchemaTask should download specific schema version`() {
         buildFile = folderRule.newFile("build.gradle")
         buildFile.writeText(
             """
@@ -231,7 +125,8 @@ class DownloadTaskIT : TestContainersUtils() {
             schemaRegistry {
                 url = '$schemaRegistryEndpoint'
                 download {
-                    subject('test-subject', 'src/main/avro/test', 1)
+                    subject('$subject', 'src/main/avro/test_v1', 1)
+                    subject('$subject', 'src/main/avro/test_v2', 2)
                 }
             }
         """
@@ -245,10 +140,16 @@ class DownloadTaskIT : TestContainersUtils() {
             .withDebug(true)
             .build()
 
-        Assertions.assertThat(File(folderRule.root, "src/main/avro/test")).exists()
-        val resultFile = File(folderRule.root, "src/main/avro/test/test-subject.avsc")
-        Assertions.assertThat(resultFile).exists()
         Assertions.assertThat(result?.task(":downloadSchemasTask")?.outcome).isEqualTo(TaskOutcome.SUCCESS)
-        Assertions.assertThat(resultFile.readText()).contains("desc")
+
+        Assertions.assertThat(File(folderRule.root, "src/main/avro/test_v1")).exists()
+        val resultFile1 = File(folderRule.root, "src/main/avro/test_v1/test-subject.avsc")
+        Assertions.assertThat(resultFile1).exists()
+        Assertions.assertThat(resultFile1.readText()).doesNotContain("desc")
+
+        Assertions.assertThat(File(folderRule.root, "src/main/avro/test_v2")).exists()
+        val resultFile2 = File(folderRule.root, "src/main/avro/test_v2/test-subject.avsc")
+        Assertions.assertThat(resultFile2).exists()
+        Assertions.assertThat(resultFile2.readText()).contains("desc")
     }
 }
